refactor the lib to module
