+++ src/device/mod.rs
@@ -1,0 +1,5 @@
+/// The `G29`` module represents the G29 device and provides methods for controlling and interacting with it.
+pub mod g29;
+
+/// The `G29Driver`` module is the underlying driver for the G29 device, managing communication and state.
+pub mod g29_driver;
